+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# 1.0
+# 1.0.0
 ## Versioncode 1
 * First release
-
+# 1.1.0
+## Versioncode 2
+* Fixed crashes on Android 9.0
+# 1.1.1
+## Versioncode 3
+* Fixed changelog dialog button visibility